backend/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
alpha-vantage==2.3.1
annotated-types==0.6.0
APScheduler==3.10.4
argcomplete==3.2.3
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
av==12.0.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloudpickle==3.0.0
cloup==2.1.2
colorama==0.4.6
colorlog==6.8.2
colour==0.1.5
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.50.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.4
glcontext==2.5.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h5py==3.10.0
html5lib==1.1
huggingface-hub==0.23.4
idna==3.6
imageio==2.34.0
intel-openmp==2021.4.0
ipython==8.22.2
isosurfaces==0.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
keras==3.1.1
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
lazy_loader==0.3
libclang==18.1.1
llvmlite==0.43.0
lxml==5.1.0
Mako==1.3.2
manim==0.18.0
manim-slides==5.1.3
ManimPango==0.4.4
mapbox-earcut==1.0.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
moderngl==5.10.0
moderngl-window==2.4.4
mpmath==1.3.0
multidict==6.0.5
multipledispatch==1.0.0
multitasking==0.0.11
namex==0.0.7
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.0
optuna-integration==3.6.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
peewee==3.17.5
Pillow==9.5.0
pipx==1.4.3
platformdirs==4.2.0
plotly==5.22.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pycairo==1.26.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic_core==2.16.3
pydub==0.25.1
pygame==2.5.2
pyglet==2.0.14
Pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.1.2
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==0.6.23
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
QtPy==2.4.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rtoml==0.10.0
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
screeninfo==0.8.1
seaborn==0.13.2
setuptools==69.2.0
shap==0.46.0
six==1.16.0
skia-pathops==0.8.0.post1
slicer==0.0.8
soupsieve==2.5
SQLAlchemy==2.0.29
srt==3.5.3
stack-data==0.6.3
svgelements==1.9.6
sympy==1.12
TA-Lib==0.4.32
tbb==2021.13.0
tenacity==8.4.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.2
traitlets==5.14.2
transformers==4.42.3
twelvedata==1.2.15
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
userpath==1.9.2
validators==0.24.0
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
xgboost==2.1.0
XlsxWriter==3.2.0
yarl==1.9.4
yfinance==0.2.40
